requirements.txt
asgiref==3.8.1
astroid==2.5.3
attrs==22.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
cloudinary==1.42.2
colorama==0.4.4
Django==5.1.6
django-cors-headers==4.7.0
django-crispy-forms==1.11.2
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.9
idna==3.3
importlib_metadata==8.6.1
inflection==0.5.1
isort==5.8.0
itypes==1.2.0
Jinja2==3.1.5
jsonschema==3.2.0
jsonschema-specifications==2024.10.1
lazy-object-proxy==1.6.0
Markdown==3.7
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.3
pillow==11.1.0
psycopg2-binary==2.9.10
pylint==2.7.4
pyparsing==3.0.9
pyrsistent==0.18.0
python-dotenv==1.0.1
python-http-client==3.3.4
pytz==2021.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.26.0
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
sendgrid==6.9.3
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==2.0.3
toml==0.10.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.12.1
zipp==3.21.0
dotenv==0.9.9
